requirements.txt
alembic==0.9.1
aniso8601==1.2.1
click==6.7
Flask==0.12.1
Flask-Migrate==2.0.3
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.6
MarkupSafe==1.0
PyMySQL==0.7.11
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
pyjwt==1.5.0